# backend/src/quotes/static/swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Quotes API
  description: API для обработки котировок
paths:
...
  /api/observability:
    get:
      tags:
        - Observability
      security:
        - BearerAuth: []
      summary: Get a list of incidents
      description: Retrieve incidents with optional filters and pagination.
      parameters:
        - name: state
          in: query
          description: Filter by incident state (e.g., Active, In Progress, Resolved).
          required: false
          schema:
            type: string
        - name: priority
          in: query
          description: Filter by priority (e.g., High, Medium, Low).
          required: false
          schema:
            type: string
        - name: service
          in: query
          description: Filter by service name.
          required: false
          schema:
            type: string
        - name: trace_id
          in: query
          description: Filter by trace ID.
          required: false
          schema:
            type: integer
        - name: last_updated
          in: query
          description: Filter by incidents updated after this timestamp (ISO 8601 format).
          required: false
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Page number for pagination.
          required: false
          schema:
            type: integer
        - name: per_page
          in: query
          description: Number of records per page.
          required: false
          schema:
            type: integer
      responses:
        200:
          description: A list of incidents with pagination details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  total_pages:
                    type: integer
                  total_records:
                    type: integer
                  per_page:
                    type: integer
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id_incident:
                          type: integer
                        state:
                          type: string
                        priority:
                          type: string
                        description:
                          type: string
                          nullable: true
                        last_updated:
                          type: string
                          format: date-time
                        service:
                          type: string
                        trace_id:
                          type: integer
        400:
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    post:
      tags:
        - Observability
      security:
        - BearerAuth: []
      summary: Filter incidents
      description: Apply filters to retrieve incidents.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state:
                  type: string
                  nullable: true
                priority:
                  type: string
                  nullable: true
                service:
                  type: string
                  nullable: true
                trace_id:
                  type: integer
                  nullable: true
                last_updated:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        200:
          description: A list of incidents matching the filters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  total_pages:
                    type: integer
                  total_records:
                    type: integer
                  per_page:
                    type: integer
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id_incident:
                          type: integer
                        state:
                          type: string
                        priority:
                          type: string
                        description:
                          type: string
                          nullable: true
                        last_updated:
                          type: string
                          format: date-time
                        service:
                          type: string
                        trace_id:
                          type: integer
        400:
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/observability/create:
    post:
      tags:
        - Observability
      security:
        - BearerAuth: []
      summary: Create a new incident
      description: Add a new incident to the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - state
                - priority
                - service
                - trace_id
              properties:
                state:
                  type: string
                priority:
                  type: string
                description:
                  type: string
                  nullable: true
                service:
                  type: string
                trace_id:
                  type: integer
      responses:
        201:
          description: Incident successfully created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  incident:
                    type: integer
        400:
          description: Invalid input data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/observability/{id_incident}:
    put:
      tags:
        - Observability
      security:
        - BearerAuth: []
      summary: Update an existing incident
      description: Modify the details of an existing incident by its ID.
      parameters:
        - name: id_incident
          in: path
          required: true
          description: The ID of the incident to update.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - state
                - priority
                - service
                - trace_id
              properties:
                state:
                  type: string
                priority:
                  type: string
                description:
                  type: string
                  nullable: true
                service:
                  type: string
                trace_id:
                  type: integer
      responses:
        200:
          description: Incident successfully updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  incident:
                    type: integer
        400:
          description: Invalid input data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        404:
          description: Incident not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...